needed a bigger type for writeLenBytes
git-svn-id: http://svn.eol.ucar.edu/svn/eol/common/branches/pentek-windriver@8427 da89c675-a4d7-0310-8a15-9e7546503515

+++ p7142.cpp
@@ -469,7 +469,7 @@ int p7142::_ddrMemWrite (P7142_REG_ADDR* p7142Regs,
     P7142_DDR_MEM_PARAMS ddrMemParams; /* DDR memory params */
     unsigned int         bankDepth;
     int                  status;
-    uint16_t             writeLenBytes;
+    uint32_t             writeLenBytes;
     uint16_t             writeLenBlocks;
 
     // We need to write in 32-byte blocks, and we need to write at least
